Change setSubtitleToolbarIcon behavior
